Refactor: separate edit domain action saga
