+++ src/modules/dashboard/sagas/actions.ts
@@ -26,168 +26,6 @@ import {
 } from '../../core/actionCreators';
 import { updateColonyDisplayCache } from './utils';
 
-function* editDomainAction({
-  payload: {
-    colonyAddress,
-    colonyName,
-    domainName,
-    domainColor,
-    domainPurpose,
-    annotationMessage,
-    domainId,
-  },
-  meta: { id: metaId, history },
-  meta,
-}: Action<ActionTypes.COLONY_ACTION_DOMAIN_EDIT>) {
-  let txChannel;
-  try {
-    const apolloClient = TEMP_getContext(ContextModule.ApolloClient);
-
-    if (!domainId) {
-      throw new Error('A domain id is required to edit domain');
-    }
-
-    txChannel = yield call(getTxChannel, metaId);
-
-    const batchKey = 'editDomainAction';
-    const {
-      editDomainAction: editDomain,
-      annotateEditDomainAction: annotateEditDomain,
-    } = yield createTransactionChannels(metaId, [
-      'editDomainAction',
-      'annotateEditDomainAction',
-    ]);
-
-    const createGroupTransaction = ({ id, index }, config) =>
-      fork(createTransaction, id, {
-        ...config,
-        group: {
-          key: batchKey,
-          id: metaId,
-          index,
-        },
-      });
-
-    yield createGroupTransaction(editDomain, {
-      context: ClientType.ColonyClient,
-      methodName: 'editDomainWithProofs',
-      identifier: colonyAddress,
-      params: [],
-      ready: false,
-    });
-
-    if (annotationMessage) {
-      yield createGroupTransaction(annotateEditDomain, {
-        context: ClientType.ColonyClient,
-        methodName: 'annotateTransaction',
-        identifier: colonyAddress,
-        params: [],
-        ready: false,
-      });
-    }
-
-    yield takeFrom(editDomain.channel, ActionTypes.TRANSACTION_CREATED);
-    if (annotationMessage) {
-      yield takeFrom(
-        annotateEditDomain.channel,
-        ActionTypes.TRANSACTION_CREATED,
-      );
-    }
-
-    yield put(transactionPending(editDomain.id));
-
-    /*
-     * Upload domain metadata to IPFS
-     */
-    let domainMetadataIpfsHash = null;
-    domainMetadataIpfsHash = yield call(
-      ipfsUpload,
-      JSON.stringify({
-        domainName,
-        domainColor,
-        domainPurpose,
-      }),
-    );
-
-    yield put(
-      transactionAddParams(editDomain.id, [
-        domainId,
-        (domainMetadataIpfsHash as unknown) as string,
-      ]),
-    );
-
-    yield put(transactionReady(editDomain.id));
-
-    const {
-      payload: { hash: txHash },
-    } = yield takeFrom(
-      editDomain.channel,
-      ActionTypes.TRANSACTION_HASH_RECEIVED,
-    );
-    yield takeFrom(editDomain.channel, ActionTypes.TRANSACTION_SUCCEEDED);
-
-    if (annotationMessage) {
-      yield put(transactionPending(annotateEditDomain.id));
-
-      /*
-       * Upload annotationMessage to IPFS
-       */
-      let annotationMessageIpfsHash = null;
-      if (annotationMessage) {
-        annotationMessageIpfsHash = yield call(
-          ipfsUpload,
-          JSON.stringify({
-            annotationMessage,
-          }),
-        );
-      }
-
-      yield put(
-        transactionAddParams(annotateEditDomain.id, [
-          txHash,
-          annotationMessageIpfsHash,
-        ]),
-      );
-
-      yield put(transactionReady(annotateEditDomain.id));
-
-      yield takeFrom(
-        annotateEditDomain.channel,
-        ActionTypes.TRANSACTION_SUCCEEDED,
-      );
-    }
-
-    /*
-     * Update the colony object cache
-     */
-    yield apolloClient.query<ColonyFromNameQuery, ColonyFromNameQueryVariables>(
-      {
-        query: ColonyFromNameDocument,
-        variables: { name: colonyName || '', address: colonyAddress },
-        fetchPolicy: 'network-only',
-      },
-    );
-
-    yield put<AllActions>({
-      type: ActionTypes.COLONY_ACTION_DOMAIN_EDIT_SUCCESS,
-      meta,
-    });
-
-    if (colonyName) {
-      yield routeRedirect(`/colony/${colonyName}/tx/${txHash}`, history);
-    }
-  } catch (error) {
-    return yield putError(
-      ActionTypes.COLONY_ACTION_DOMAIN_EDIT_ERROR,
-      error,
-      meta,
-    );
-  } finally {
-    txChannel.close();
-  }
-  return null;
-}
-
 function* editColonyAction({
   payload: {
     colonyAddress,
@@ -590,7 +428,6 @@ function* colonyTokenUnlock({
 
 export default function* tasksSagas() {
   yield takeEvery(ActionTypes.COLONY_ACTION_EDIT_COLONY, editColonyAction);
-  yield takeEvery(ActionTypes.COLONY_ACTION_DOMAIN_EDIT, editDomainAction);
   yield takeEvery(
     ActionTypes.COLONY_ACTION_USER_ROLES_SET,
     createSetUserRolesAction,
