+++ src/modules/dashboard/sagas/actions/index.ts
@@ -5,6 +5,7 @@ import moveFundsActionSaga from './moveFunds';
 import mintTokensActionSaga from './mintTokens';
 import versionUpgradeActionSaga from './versionUpgrade';
 import createDomainActionSaga from './createDomain';
+import editDomainActionSaga from './editDomain';
 
 export default function* actionsSagas() {
   yield all([
@@ -13,5 +14,6 @@ export default function* actionsSagas() {
     call(mintTokensActionSaga),
     call(versionUpgradeActionSaga),
     call(createDomainActionSaga),
+    call(editDomainActionSaga),
   ]);
 }
